+++ include/system.h
@@ -14,7 +14,7 @@ class System {
   Processor& Cpu();
   std::vector<Process>& Processes();
   float MemoryUtilization();
-  long UpTime();
+  float UpTime();
   int TotalProcesses();
   int RunningProcesses();
   std::string Kernel();
